Merge dev into main

+++ chat-bot-cm-onboard-backend/app/models/pharmacy_inventory.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Text, DateTime, func, ForeignKey
+from sqlalchemy import Column, Integer, String, Text, DateTime, func, ForeignKey, DECIMAL
 from app.core.database import Base
 
 
@@ -9,7 +9,7 @@ class PharmacyInventory(Base):
     drug_name = Column(String(100), nullable=False, index=True)
     category = Column(String(50), nullable=False, index=True)
     brand = Column(String(50), nullable=False, index=True)
-    unit_price = Column(Integer, nullable=False)
+    unit_price = Column(DECIMAL(10,2), nullable=False)
     quantity_in_stock = Column(Integer, nullable=False)
     expiry_date = Column(DateTime, nullable=False)
     supplier_id = Column(String(50), ForeignKey("suppliers.supplier_id"), nullable=False, index=True)
